requirements.txt
python-socketio[client]
numpy
requests